Enables entity setting and config reloading
Tweaks to config.yml to be more cave-y.

Entities are set to not despawn. Probabilities of materials and entities
can have decimals and be sub 1% if required. For some entities, 1% was
too high a rate.

Adds a debug setting in config.yml that will enable you to see where
entities have been spawned. It can be removed at some point.