Fix issue when alternative teleports did not  work. (#4).
There was a mistake with checking if position is changed in Y coordinate. It took from position and compare it to from position, so always fails.
This issue has been fixed now.

Also to improve ability to change these settings on the fly, I add them as world protection flags, so admins can enable and disable them via config.